+++ port/cortex-m0/src/interrupt/pending.rs
@@ -12,7 +12,7 @@ pub struct PendingControl {
 }
 
 impl PendingControl {
-  pub fn new(base_addr: u32) -> Self {
+  pub fn new(base_addr: usize) -> Self {
     PendingControl {
       ispr: ISPR::new(base_addr),
       icpr: ICPR::new(base_addr),
@@ -34,19 +34,19 @@ impl PendingControl {
 
 #[derive(Copy, Clone)]
 struct ISPR {
-  base_addr: u32,
+  base_addr: usize,
 }
 
 impl Register for ISPR {
-  fn new(base_addr: u32) -> Self {
+  fn new(base_addr: usize) -> Self {
     ISPR { base_addr: base_addr }
   }
 
-  fn base_addr(&self) -> u32 {
+  fn base_addr(&self) -> usize {
     self.base_addr
   }
   
-  fn mem_offset(&self) -> u32 {
+  fn mem_offset(&self) -> usize {
     0x100
   }
 }
@@ -78,19 +78,19 @@ impl ISPR {
 
 #[derive(Copy, Clone)]
 struct ICPR {
-  base_addr: u32,
+  base_addr: usize,
 }
 
 impl Register for ICPR {
-  fn new(base_addr: u32) -> Self {
+  fn new(base_addr: usize) -> Self {
     ICPR { base_addr: base_addr }
   }
 
-  fn base_addr(&self) -> u32 {
+  fn base_addr(&self) -> usize {
     self.base_addr
   }
 
-  fn mem_offset(&self) -> u32 {
+  fn mem_offset(&self) -> usize {
     0x180
   }
 }
